move tx LSM LS
