+++ x/stakeibc/client/cli/tx.go
@@ -82,3 +82,37 @@ func CmdLiquidStake() *cobra.Command {
 
 	return cmd
 }
+
+func CmdLSMLiquidStake() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "lsm-liquid-stake [amount] [lsm-token-denom]",
+		Short: "Broadcast message lsm-liquid-stake",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			amount, found := sdk.NewIntFromString(args[0])
+			if !found {
+				return errorsmod.Wrap(sdkerrors.ErrInvalidType, "can not convert string to int")
+			}
+			denom := args[1]
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgLSMLiquidStake(
+				clientCtx.GetFromAddress().String(),
+				amount,
+				denom,
+			)
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
