+++ x/stakeibc/client/cli/tx_lsm_liquid_stake.go
@@ -1,48 +1,1 @@
 package cli
-
-import (
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/client/tx"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	errorsmod "cosmossdk.io/errors"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/spf13/cobra"
-
-	"github.com/Stride-Labs/stride/v22/x/stakeibc/types"
-)
-
-func CmdLSMLiquidStake() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "lsm-liquid-stake [amount] [lsm-token-denom]",
-		Short: "Broadcast message lsm-liquid-stake",
-		Args:  cobra.ExactArgs(2),
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			amount, found := sdk.NewIntFromString(args[0])
-			if !found {
-				return errorsmod.Wrap(sdkerrors.ErrInvalidType, "can not convert string to int")
-			}
-			denom := args[1]
-
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			msg := types.NewMsgLSMLiquidStake(
-				clientCtx.GetFromAddress().String(),
-				amount,
-				denom,
-			)
-			if err := msg.ValidateBasic(); err != nil {
-				return err
-			}
-			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
-		},
-	}
-
-	flags.AddTxFlagsToCmd(cmd)
-
-	return cmd
-}
